up docker thanh cong

+++ database/migrations/2026_04_19_165610_add_note_password_to_notes_table.php
@@ -11,8 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('notes', function (Blueprint $table) {
-            $table->string('note_password')->nullable()->after('content');
+       Schema::table('notes', function (Blueprint $table) {
+            if (!Schema::hasColumn('notes', 'note_password')) {
+                $table->string('note_password')->nullable()->after('content');
+            }
         });
     }
 
@@ -22,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('notes', function (Blueprint $table) {
-            $table->dropColumn('note_password');
+            if (Schema::hasColumn('notes', 'note_password')) {
+                $table->dropColumn('note_password');
+            }
         });
     }
 };
